幸存者模式 新增UI-Effect插件

+++ Assets/Scripts/HotFix/_TopDownEngine/Characters/Exp/ExpData.cs
@@ -14,11 +14,11 @@ namespace MoreMountains
         {
             public string Key;
             public int MaxLevel = 18;
-            public float StartXpRequired = 100;
-            public float MaxLevelXpRequired = 300;
+            public int StartXpRequired = 100;
+            public int MaxLevelXpRequired = 300;
             public AnimationCurve XpCurve;
 
-            public void Deconstruct(out string key, out int maxLevel, out float startXpRequired, out float maxLevelXpRequired, out AnimationCurve xpCurve)
+            public void Deconstruct(out string key, out int maxLevel, out int startXpRequired, out int maxLevelXpRequired, out AnimationCurve xpCurve)
             {
                 key = Key;
                 maxLevel = MaxLevel;
@@ -27,7 +27,7 @@ namespace MoreMountains
                 xpCurve = XpCurve;
             }
 
-            public void Deconstruct(out int maxLevel, out float startXpRequired, out float maxLevelXpRequired, out AnimationCurve xpCurve)
+            public void Deconstruct(out int maxLevel, out int startXpRequired, out int maxLevelXpRequired, out AnimationCurve xpCurve)
             {
                 maxLevel = MaxLevel;
                 startXpRequired = StartXpRequired;
